♻️  도메인 들어가지 않는 디자인시스템 네이밍으로 변경 및 주석방식 컨벤션 적용
